% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtrix.R
\name{corr.test2apa}
\alias{corr.test2apa}
\title{Report output of cor.test in APA style}
\usage{
corr.test2apa(x, stars)
}
\arguments{
\item{stars}{Optional argument. logical = TRUE if stars are required for p-values}

\item{model}{Output of corr.test function}
}
\value{
A character vector, e.g. "r(12) = 0.12, p = 0.123"
}
\description{
Report output of cor.test in APA style
}
