% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtrix.R
\name{p2apa_plus_stars}
\alias{p2apa_plus_stars}
\title{Report p-value in APA style and add stars}
\usage{
p2apa_plus_stars(x)
}
\arguments{
\item{P-value}{Numeric}
}
\value{
A character vector, e.g. "p = 0.023*"
}
\description{
Uses "p < 0.001" if p-value less than 0.001. Otherwise outputs "p = 0.123" (note 3 decimal places)
}
