% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nixtrix.R
\name{p2apa}
\alias{p2apa}
\title{Report p-value in APA style}
\usage{
p2apa(x)
}
\arguments{
\item{p-value}{Numeric}
}
\value{
A character vector, e.g. "p = 0.123"
}
\description{
Uses "p < 0.001" if p-value less than 0.001. Otherwise outputs "p = 0.123" (note 3 decimal places)
}
